add config state for question editor
